requirements.txt
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Faker==33.1.0
pytest==8.3.4
